public int maxCapacity >> decimal

+++ SlaughterHouse/SlaughterHouse/Entities/Tray.cs
@@ -4,7 +4,7 @@ public class Tray
 {
     public int TrayId { get; set; }
     public PartType partType { get; set; }
-    public int maxCapacity { get; set; }
+    public decimal maxCapacity { get; set; }
     public decimal currentWeight { get; set; }
 
     public Tray(int trayId, PartType partType, int maxCapacity)
